AP-5689 Improvement in Submenu
